refactor : feature user 폴더명 변경
