+++ src/widgets/post/ui/PostsTableContainer.tsx
@@ -1,6 +1,6 @@
 import { useQueryPostList } from "@features/post/api/api"
 import { PostSearchFilters } from "@features/post/ui/PostSearchFilters"
-import { useUserCache } from "@features/user-management/api/api"
+import { useUserCache } from "@features/user/api/api"
 import { PostTotalTable } from "./PostTotalTable"
 import { PostPagination } from "./PostPagination"
